pyproject.toml
[tool.poetry]
name = "usd-rub"
version = "0.1.0"
description = ""
authors = ["Olonov Aleksandr <adrolv08111985@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
environs = "^11.0.0"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
